requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==3.0.0
WTForms==3.1.0